Code Improvement: `replace` instead of `navigate`
We can use `navigation.replace("Dashboard")` instead
of `navigation.navigate("Dashboard")` to replace
the current route in the history stack. This way
we don't need to add the `headerLeft: () => null`
in the `Dashboard` screen to prevent the user from
going back to the login screen after logging in
with the back button on the app bar.

We can now safely remove the `headerLeft` prop from
the `Dashboard` screen. This way we have a cleaner
code. The app after this change has been tested on
both Android and iOS, everything works as expected.